requirements.txt
dispatch==2.0.5
Django==3.2.5
wagtail==2.14.1
git+git://github.com/ubyssey/django-google-storage.git@v0.5.4#egg=django-google-storage-updated
django-storages[google]==1.10
django-canonical-domain==0.3.0
pymemcache==3.5.0
mysqlclient==2.0.1
django-phonenumber-field==2.0.0
phonenumbers==8.12.9
beautifulsoup4==4.8.2
feedparser==6.0.1
pyyaml==5.4
ua-parser==0.10.0
user-agents==2.2.0
django-user_agents==0.4.0
gunicorn==20.0.4
Pillow==8.2.0
google-cloud-datastore==1.15.0
google-cloud-storage==1.31.0
google-cloud-secret-manager==1.0.0
django-environ==0.4.5
requests==2.24.0
whitenoise==5.1.0
brotli==1.0.9
django-extensions==3.1.3
git+git://github.com/wamberg/django-dbtemplates.git@272af3c#egg=django-dbtemplates
wagtail-modelchooser==2.13.1
wagtailmenus==3.0.2
django-debug-toolbar==3.2.1
pylint==2.8.3
pylint-django==2.4.4
ipython